modified:   .cproject 	modified:   .settings/stm32cubeide.project.prefs 	modified:   Core/Inc/tables.h 	modified:   Core/Src/main.c 	modified:   Core/Src/stm32f1xx_it.c 	modified:   Core/Src/tables.c 	new file:   Debug.launch 	modified:   Drivers/CMSIS/DSP/Include/arm_common_tables.h 	modified:   Drivers/CMSIS/DSP/Source/CommonTables/arm_common_tables.c 	modified:   Drivers/CMSIS/DSP/Source/TransformFunctions/arm_rfft_init_q31.c 	modified:   Drivers/CMSIS/DSP/Source/TransformFunctions/arm_rfft_q31.c 	renamed:    fft Release.launch -> Release.launch 	modified:   fft.ioc

+++ Core/Inc/tables.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-extern const int16_t expMap[];
-extern const int16_t blackmanHarris1024[];
+extern const int32_t expMap[];
+extern const int32_t blackmanHarris1024[];
 
 #endif // _TABLES_H_
